front: move train deletion feature from TimetableToolbar to BoardWrapper menu
